feat: implement edit and delete expense functionality
- Add EditExpenseView with modern styling matching AddExpenseView
- Implement updateExpense and deleteExpense methods in ExpensesStore
- Add database operations for updating and deleting expenses
- Convert edit form from sheet to navigation for better performance
- Move edit/delete buttons from nav bar to expense detail header
- Fix automatic refresh of expense detail view after updates
- Update styling to use modern toolbar API and remove deprecated navigationBarItems
- Fix UUID conversion issues and optional handling
- Remove backup folder and clean up duplicate files
- Update balance colors to match title styling (remove green/orange from totals)

Key improvements:
- Better UX with prominent action buttons in header
- Automatic UI updates without manual refresh
- Modern iOS styling and navigation patterns
- Robust error handling and validation
- Performance optimizations
